refactor: inline to footer menu and changed base item class name
