+++ components/02-molecules/menu/_menu-list.twig
@@ -3,7 +3,7 @@
 {% set ul_blockname = menu_blockname %}
 
 {% set item_blockname = menu_blockname %}
-{% set item_base_class = 'item' %}
+{% set item_base_class = 'menu-item' %}
 
 {# List #}
 {% extends "@atoms/lists/ul.twig" %}
