+++ components/02-molecules/menu/menu.yml
@@ -1,4 +1,4 @@
-menu_items:
+items:
   - title: 'Menu Item 1'
     url: '#'
     below:
