+++ components/03-organisms/menus/footer/footer-menu.twig
@@ -1,0 +1,4 @@
+{% include "@molecules/menu/menu.twig" with {
+  menu_class: 'footer-menu',
+  items: footer_menu_items,
+} %}
