+++ components/03-organisms/menus/footer/footer-menu.yml
@@ -1,4 +1,4 @@
-inline_menu_items:
+footer_menu_items:
   - title: 'Test'
     url: '#'
   - title: 'Number 2'
