+++ components/03-organisms/menus/menus.stories.js
@@ -1,10 +1,10 @@
 import breadcrumb from './breadcrumbs/breadcrumbs.twig';
-import inlineMenu from './inline/inline-menu.twig';
+import footerMenu from './footer/footer-menu.twig';
 import mainMenu from './main-menu/main-menu.twig';
 import socialMenu from './social/social-menu.twig';
 
 import breadcrumbsData from './breadcrumbs/breadcrumbs.yml';
-import inlineMenuData from './inline/inline-menu.yml';
+import footerMenuData from './footer/footer-menu.yml';
 import mainMenuData from './main-menu/main-menu.yml';
 import socialMenuData from './social/social-menu.yml';
 
@@ -13,11 +13,11 @@ import './main-menu/main-menu';
 /**
  * Storybook Definition.
  */
-export default { title: 'Molecules/Menus' };
+export default { title: 'Organisms/Menus' };
 
 export const breadcrumbs = () => breadcrumb(breadcrumbsData);
 
-export const inline = () => inlineMenu(inlineMenuData);
+export const footer = () => footerMenu(footerMenuData);
 
 export const main = () => mainMenu(mainMenuData);
