+++ components/03-organisms/site/site-footer/site-footer.twig
@@ -9,7 +9,7 @@
     </div>
     <div {{ bem('menu', [], footer__base_class) }}>
       {% block footer__menu %}
-        {% include "@organisms/menus/inline/inline-menu.twig" %}
+        {% include "@organisms/menus/footer/footer-menu.twig" %}
       {% endblock %}
     </div>
   </div>
