+++ components/05-pages/landing-pages/landing-pages.stories.js
@@ -5,7 +5,7 @@ import home from './home.twig';
 import mainMenuData from '../../03-organisms/menus/main-menu/main-menu.yml';
 import breadcrumbData from '../../03-organisms/menus/breadcrumbs/breadcrumbs.yml';
 import socialMenuData from '../../03-organisms/menus/social/social-menu.yml';
-import footerMenuData from '../../03-organisms/menus/inline/inline-menu.yml';
+import footerMenuData from '../../03-organisms/menus/footer/footer-menu.yml';
 
 /**
  * Storybook Definition.
